+++ src/barril/units/_abstractvaluewithquantity.py
@@ -7,7 +7,7 @@ from barril.units.unit_database import UnitDatabase
 
 from ._quantity import ObtainQuantity
 
-__all__ = ["AbstractValueWithQuantityObject"]
+__all__ = [str("AbstractValueWithQuantityObject")]  # pylint: disable=invalid-all-object
 
 
 class AbstractValueWithQuantityObject(object):