+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 CHANGELOG
 =========
 
+1.1.0 (2018-24-2018)
+--------------------
+
+* Add ``number`` and ``fraction`` properties to ``FractionValue``.
+* Add ``unit`` read-only property to ``Quantity``.
+
 1.0.0 (2018-21-2018)
 --------------------